[MediaStoreUtil] read dataAdded from MediaStore

+++ app/src/main/java/player/phonograph/provider/MusicPlaybackQueueStore.java
@@ -22,6 +22,7 @@ import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.provider.BaseColumns;
 import android.provider.MediaStore.Audio.AudioColumns;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
@@ -32,15 +33,15 @@ import java.util.List;
 
 /**
  * @author Andrew Neal, modified for Phonograph by Karim Abou Zeid
- *         <p/>
- *         This keeps track of the music playback and history state of the playback service
+ * <p/>
+ * This keeps track of the music playback and history state of the playback service
  */
 public class MusicPlaybackQueueStore extends SQLiteOpenHelper {
     @Nullable
     private static MusicPlaybackQueueStore sInstance = null;
     public static final String PLAYING_QUEUE_TABLE_NAME = "playing_queue";
     public static final String ORIGINAL_PLAYING_QUEUE_TABLE_NAME = "original_playing_queue";
-    private static final int VERSION = 4;
+    private static final int VERSION = 5;
 
     /**
      * Constructor of <code>MusicPlaybackState</code>
@@ -82,6 +83,9 @@ public class MusicPlaybackQueueStore extends SQLiteOpenHelper {
         builder.append(AudioColumns.DATA);
         builder.append(" STRING NOT NULL,");
 
+        builder.append(AudioColumns.DATE_ADDED);
+        builder.append(" LONG NOT NULL,");
+
         builder.append(AudioColumns.DATE_MODIFIED);
         builder.append(" LONG NOT NULL,");
 
@@ -102,10 +106,19 @@ public class MusicPlaybackQueueStore extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(@NonNull final SQLiteDatabase db, final int oldVersion, final int newVersion) {
-        // not necessary yet
-        db.execSQL("DROP TABLE IF EXISTS " + PLAYING_QUEUE_TABLE_NAME);
-        db.execSQL("DROP TABLE IF EXISTS " + ORIGINAL_PLAYING_QUEUE_TABLE_NAME);
-        onCreate(db);
+        if (oldVersion == 4 && newVersion == 5) {
+            migrate4to5(db, oldVersion, newVersion);
+        } else {
+            // not necessary yet
+            db.execSQL("DROP TABLE IF EXISTS " + PLAYING_QUEUE_TABLE_NAME);
+            db.execSQL("DROP TABLE IF EXISTS " + ORIGINAL_PLAYING_QUEUE_TABLE_NAME);
+            onCreate(db);
+        }
+    }
+
+    private void migrate4to5(@NonNull final SQLiteDatabase db, final int oldVersion, final int newVersion) {
+        db.execSQL("ALTER TABLE "+ PLAYING_QUEUE_TABLE_NAME + " ADD " + AudioColumns.DATE_ADDED + " LONG NOT NULL DEFAULT 0");
+        db.execSQL("ALTER TABLE "+ ORIGINAL_PLAYING_QUEUE_TABLE_NAME + " ADD " + AudioColumns.DATE_ADDED + " LONG NOT NULL DEFAULT 0");
     }
 
     @Override
@@ -165,6 +178,7 @@ public class MusicPlaybackQueueStore extends SQLiteOpenHelper {
                     values.put(AudioColumns.YEAR, song.year);
                     values.put(AudioColumns.DURATION, song.duration);
                     values.put(AudioColumns.DATA, song.data);
+                    values.put(AudioColumns.DATE_ADDED, song.dateAdded);
                     values.put(AudioColumns.DATE_MODIFIED, song.dateModified);
                     values.put(AudioColumns.ALBUM_ID, song.albumId);
                     values.put(AudioColumns.ALBUM, song.albumName);
